Move ModuleName enum outside of arch package, because it will have to be interacted with often, unlike the rest of the arch
