+++ src/main/java/com/worldwarofants/game/arch/AbstractController.java
@@ -2,7 +2,7 @@ package com.worldwarofants.game.arch;
 
 import com.worldwarofants.game.arch.console.command.ICommand;
 import com.worldwarofants.game.arch.module.IModuleNavigator;
-import com.worldwarofants.game.arch.module.ModuleName;
+import com.worldwarofants.game.module.ModuleName;
 import com.worldwarofants.game.model.World;
 
 /**
