+++ src/main/java/com/worldwarofants/game/arch/console/command/AbstractCommandHandler.java
@@ -3,7 +3,7 @@ package com.worldwarofants.game.arch.console.command;
 import com.worldwarofants.game.arch.AbstractController;
 import com.worldwarofants.game.arch.module.AbstractModule;
 import com.worldwarofants.game.arch.module.IModuleNavigator;
-import com.worldwarofants.game.arch.module.ModuleName;
+import com.worldwarofants.game.module.ModuleName;
 
 import java.util.Arrays;
 import java.util.Comparator;
